Basic Functionality Complete
This is the last push to basic functionality being complete. An account is able to be created and logged in with a username and display name. From this games are able to be viewed and entered. The user can add other accounts based off username and view other's profiles by clicking on them. Profiles current just display recent games and friends, but past basic functionality they will have favorite games listed at the bottom and other things appropriate for a profile page.